refactor: 그룹 퀘스트 관련 응용 클래스를 분리한다
